Accept float crop data for easier web integration

+++ Services/PhotoProcessing/CropTransform.cs
@@ -2,9 +2,9 @@
 {
     public class CropTransform
     {
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float Width { get; set; }
+        public float Height { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
     }
 }
